post game jam

+++ include/EWEngine/Data/EngineDataTypes.h
@@ -8,4 +8,4 @@ typedef uint16_t TextureID;
 typedef uint16_t TransformID;
 typedef uint16_t Compute_TextureID;
 typedef uint32_t SkeletonID;
-typedef uint16_t PipelineID;
+typedef uint32_t PipelineID;
